+++ CHANGELOG.md
@@ -7,8 +7,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Évolution
 - Redmine 26581 : Intégration conditionnelle de Sentry et refonte du logging
-  - Nouvelles variables d'environnement : `SENTRY_ENABLED`, `SENTRY_DSN`
-  - Initialisation de Sentry uniquement si activé et DSN défini
+  - Nouvelles variables d'environnement : `ENV_MODE`, `SENTRY_DSN`
+  - Initialisation de Sentry uniquement si DSN défini
   - Ajustement du logging pour éviter les doublons et limiter l’envoi à Sentry aux erreurs uniquement
 
 - Redmine 27831 : Amélioration de la recherche textuelle dans les listes de valeurs pré-enregistrées